+++ lib/src/main/kotlin/com/ithersta/tgbotapi/fsm/BaseStatefulContext.kt
@@ -3,6 +3,7 @@ package com.ithersta.tgbotapi.fsm
 import dev.inmo.tgbotapi.bot.RequestsExecutor
 import kotlinx.coroutines.CoroutineScope
 
+@FsmDsl
 interface BaseStatefulContext<BS : Any, BU : Any, S : BS, U : BU> : RequestsExecutor {
     val state: S
     val setState: suspend (BS) -> Unit