+++ lib/src/main/kotlin/com/ithersta/tgbotapi/fsm/builders/NestedStateMachineBuilder.kt
@@ -1,13 +1,15 @@
 package com.ithersta.tgbotapi.fsm.builders
 
 import com.ithersta.tgbotapi.fsm.FsmDsl
-import com.ithersta.tgbotapi.fsm.entities.RoleFilter
+import com.ithersta.tgbotapi.fsm.entities.NestedStateMachine
 import com.ithersta.tgbotapi.fsm.entities.StateFilter
+import com.ithersta.tgbotapi.fsm.entities.StateMachine
 import org.koin.core.component.KoinComponent
 
 @FsmDsl
-class RoleFilterBuilder<BS : Any, BU : Any, U : BU, K : Any>(
-    private val map: (BU) -> U?
+class NestedStateMachineBuilder<BS : Any, BU : Any, S : BS, U : BU, K : Any, R : Any>(
+    private val level: Int,
+    private val onExit: S.(R) -> BS
 ) : KoinComponent {
     private val filters = mutableListOf<StateFilter<BS, BU, *, U, K>>()
 
@@ -20,10 +22,12 @@ class RoleFilterBuilder<BS : Any, BU : Any, U : BU, K : Any>(
     }
 
     fun <S : BS> state(block: StateFilterBuilder<BS, BU, S, U, K>.() -> Unit, map: (BS) -> S?) {
-        filters += StateFilterBuilder<BS, BU, S, U, K>(map, 0).apply(block).build()
+        filters += StateFilterBuilder<BS, BU, S, U, K>(map, level).apply(block).build()
     }
 
-    fun build(): RoleFilter<BS, BU, U, K> {
-        return RoleFilter(map, filters)
+    suspend fun exit(state: StateMachine<BS, *, *>.StateHolder<*>, result: R) {
+        state.popAndOverride { onExit(it as S, result) }
     }
+
+    fun build() = NestedStateMachine(level, filters)
 }