+++ lib/src/main/kotlin/com/ithersta/tgbotapi/fsm/repository/StateRepository.kt
@@ -1,6 +1,6 @@
 package com.ithersta.tgbotapi.fsm.repository
 
 interface StateRepository<Key : Any, BaseState : Any> {
-    fun get(key: Key): BaseState
-    fun set(key: Key, state: BaseState)
+    fun get(key: Key): List<BaseState>?
+    fun set(key: Key, stateStack: List<BaseState>)
 }