+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 ### Changed
 - Deprecate `onTransition`, replace with `onEnter`.
 - Make it impossible to nest triggers.
+- Artifact id is now just `fsm`